Support server-generated MQTT credentials, client-side folder name configuration
To support switching over to JWT authentication for Mosquitto, add a config option
that makes the client request its MQTT credentials via an HTTP API endpoint.

To support clients that are associated with organizations but need to report
device data for sites within the organization, allow the resource path prefix to
be set in configuration instead of fetched from the server.

The default behavior remains unchanged.